Adds __hash__ to PublicKey and TxId
PublicKey and TxId need __hash__ so they can be used as keys to Python dicts, which the bindings do. Other primitives may benefit also from this, but I', not completely sure if it's necessary to be done manually or it is only required as long as the bindings return such a map.